requirements.txt
asgiref==3.5.2
attrs==22.2.0
Babel==2.11.0
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-money==3.0.0
django-reverse-admin==2.9.6
djangorestframework==3.14.0
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==15.3.4
idna==3.4
importlib-metadata==5.1.0
inflection==0.5.1
iniconfig==1.1.1
Markdown==3.4.1
marshmallow==3.19.0
marshmallow-dataclass==8.5.10
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==22.0
pathspec==0.10.2
platformdirs==2.5.4
pluggy==1.0.0
psycopg2-binary==2.9.5
py-moneyed==2.0
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pytest==7.2.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
typing-inspect==0.8.0
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.13
zipp==3.10.0
